pyproject.toml
[project]
name = "vector-tool-search-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "TerminalMan", email = "84923604+SecretiveShell@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "llama-cpp-python>=0.3.6",
    "loopgpt>=0.1.1",
]

[project.scripts]
vector-tool-search-demo = "vector_tool_search_demo:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
